Moving clifford utils for cdr to own files.
