+++ mitiq/cdr/__init__.py
@@ -28,8 +28,6 @@ from mitiq.cdr.data_regression import (
     linear_fit_function_no_intercept,
 )
 from mitiq.cdr.clifford_training_data import (
-    count_non_cliffords,
-    is_clifford,
     generate_training_circuits,
 )
 from mitiq.cdr.cdr import execute_with_cdr, mitigate_executor, cdr_decorator
